Major: Updated install script to be more modular by linking autostart and crontab scripts to the files in this repo instead of copying those files from this repo

+++ install.sh
@@ -2,20 +2,40 @@
 # Run this script to install rpi-chrome-display to this computer (Raspberry Pi)
 # See README.md file in rpi-chrome-display project for more instructions
 
-echo 'Installing xdotool and unclutter (if not already installed)...'
-sudo apt update && sudo apt install xdotool unclutter -y
-
-# Make sure this project's autostart snippet is in the user's autostart file
-if grep -F 'chromium' ~/.config/lxsession/LXDE-pi/autostart >/dev/null 2>&1; then
-	echo "Your ~/.config/lxsession/LXDE-pi/autostart file appears to already"
-	echo " contain instructions to start Chromium when you log in."
-	echo "Double-check your autostart file to make sure it contains everything"
-	echo " in this project's autostart.part file."
-else
-	echo 'Backing up ~/.config/lxsession/LXDE-pi/autostart file...'
-	cp ~/.config/lxsession/LXDE-pi/autostart ~/.config/lxsession/LXDE-pi/autostart.bak
-	echo 'Modifying ~/.config/lxsession/LXDE-pi/autostart file...'
-	cat autostart.part >> ~/.config/lxsession/LXDE-pi/autostart
+if ! ( dpkg -s xdotool >/dev/null 2>&1 || dpkg -s unclutter >/dev/null 2>&1 ); then
+	echo 'Installing xdotool and unclutter...'
+	sudo apt update && sudo apt install xdotool unclutter -y
 fi
+
+# autostart
+if grep -F 'autostart.sh' ~/.config/lxsession/LXDE-pi/autostart >/dev/null 2>&1; then
+	echo "Your ~/.config/lxsession/LXDE-pi/autostart file appears to already"
+	echo "contain the autostart.sh script from this project."
+	echo "Double-check your autostart file to ensure it includes the following line:"
+	echo
+	echo "  @$PWD/autostart.sh"
+	echo
+else
+	if [ -e "~/.config/lssession/LXDE-pi/autostart" ]; then
+		echo 'Backing up ~/.config/lxsession/LXDE-pi/autostart file...'
+		cp ~/.config/lxsession/LXDE-pi/autostart ~/.config/lxsession/LXDE-pi/autostart.bak
+	fi
+	mkdir -p ~/.config/lxsession/LXDE-pi
+	echo 'Modifying ~/.config/lxsession/LXDE-pi/autostart file...'
+	echo "@$PWD/autostart.sh" >> ~/.config/lxsession/LXDE-pi/autostart
+fi
+
+# crontab
+	if crontab -l | grep -F 'refresh-chromium.sh' >/dev/null 2>&1; then
+		echo "Your crontab appears to already have the refresh-chromium.sh script in it."
+		echo "Double-check your crontab file by running 'crontab -e'"
+	else
+		if crontab -l >/dev/null 2>&1; then
+			echo "Appending autorefresh-chromium.cron to end of your crontab"
+		else
+			echo "Creating your crontab file using autorefresh-chromium.cron"
+		fi
+		( crontab -l ; cat autorefresh-chromium.cron ) | crontab -
+	fi
 
 echo "Installed. Follow the other steps in this project's README file and then reboot the system."
